destroy a subscriber by email, must destroy all the subscribers associated to that club.

+++ lib/sac_exact_target/models/prospect_model.rb
@@ -16,9 +16,12 @@ module SacExactTarget
     end
 
     def self.destroy_by_email(email, club_id)
-      subscriber = find_by_email(email, club_id)
-      if not subscriber.nil? and not (prospect = email_belongs_to_prospect?(subscriber.subscriber_key)).nil?
-        prospect.exact_target_prospect.destroy!
+      subscribers = find_all_by_email(email, club_id)
+      unless subscribers.empty?
+        subscribers.each do |subscriber|
+          prospect = email_belongs_to_prospect?(subscriber.subscriber_key)
+          prospect.exact_target_prospect.destroy! unless prospect.nil?
+        end
       end
     end
 
@@ -30,12 +33,16 @@ module SacExactTarget
       SacExactTarget::report_error("SacExactTarget:Prospect:destroy", res)
     end
 
-    def self.find_by_email(email, club_id)
+    def self.find_all_by_email(email, club_id)
       # Find by email . I didnt have luck looking for a subscriber by email and List.
       res = ExactTargetSDK::Subscriber.find [ ["EmailAddress", ExactTargetSDK::SimpleOperator::EQUALS, email] ]
       res.Results.collect do |result|
         result.attributes.select {|d| d == { :name => "Club", :value => club_id } }.empty? ? nil : result
-      end.flatten.first
+      end.flatten
+    end
+
+    def self.find_by_email(email, club_id)
+      find_by_email(email, club_id).first
     end
  
     # easy way to know if a subscriber key is a prospect or member. This method can be improved, filtering by List id.
